Extract bandana to separate repository
This project is extracted from an internal equinor repository. The
bandana project is a proof of concept of role based graph level access
of data to the apache jena fuseki triple store. Being a proof of
conecept the projects lacks documentation and contains some loose ends
and bugs. This will hopefully be cleaned up in upcoming commits.

Further work:

- Queries on Named graphs including the union graph urn:x-arq:UnionGraph does not work
- Update endpoints is not handled
- Code depends on scopes in the record ontology https://github.com/equinor/records. This should be abstracted to allow
  other jwt based authorization.
- Cleanup code, trailing whitespaced, prints etc.
- Improved tests
- Readme file and other documentation is needed

Co-authored-by: Eirik Nordstrand <89242375+einorequinor@users.noreply.github.com>